+++ internals/shared/src/index.ts
@@ -1,4 +1,4 @@
 export type * from './types.ts'
-export { KUBB_CONFIG_FILENAME, initDefaults, availablePlugins, pluginDefaultConfigs } from './constants.ts'
+export { KUBB_CONFIG_FILENAME, initDefaults, availablePlugins } from './constants.ts'
 export { generateConfigFile } from './init.ts'
 export { createModuleLoader } from './loader.ts'
