+++ internals/utils/src/cli/defineCommand.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from 'vitest'
-import { defineCommand } from './define.ts'
+import { defineCommand } from './defineCommand.ts'
 
 describe('defineCommand', () => {
   it('returns a command without run for parent commands (e.g. agent with subCommands)', () => {
