+++ internals/utils/src/cli/defineCommand.ts
@@ -1,4 +1,56 @@
-import type { CLIAdapter, CommandDefinition, OptionDefinition } from './types.ts'
+type BaseOption = {
+  /**
+   * Single-character short alias, e.g. `'c'` for `--config`.
+   */
+  short?: string
+  description: string
+  hint?: string
+  required?: boolean
+  /**
+   * Allowed values for string options. Used in help output and AI/MCP tool schemas.
+   */
+  enum?: Array<string>
+}
+
+type StringOption = BaseOption & { type: 'string'; default?: string }
+type BooleanOption = BaseOption & { type: 'boolean'; default?: boolean }
+
+/**
+ * A single CLI option — either a string value or a boolean flag.
+ */
+export type OptionDefinition = StringOption | BooleanOption
+
+/**
+ * The primitive type of a CLI option value.
+ */
+export type OptionType = OptionDefinition['type']
+
+/**
+ * Parsed result of `parseArgs` — typed values map plus positional arguments.
+ */
+export type ParsedArgs = {
+  values: Record<string, string | boolean | undefined>
+  positionals: Array<string>
+}
+
+/**
+ * Full definition of a CLI command including its options, sub-commands, and run handler.
+ */
+export type CommandDefinition = {
+  name: string
+  description: string
+  /**
+   * Positional argument labels shown in usage line, e.g. `['[input]']`.
+   */
+  arguments?: Array<string>
+  /**
+   * Usage examples shown in help output and exposed to AI/MCP tools.
+   */
+  examples?: Array<string>
+  options?: Record<string, OptionDefinition>
+  subCommands?: Array<CommandDefinition>
+  run?: (args: ParsedArgs) => Promise<void>
+}
 
 type OptionTypeMap = { string: string; boolean: boolean }
 
@@ -11,19 +63,6 @@ type InferValues<O extends Record<string, OptionDefinition>> = {
   [K in keyof O as IsRequired<O[K]> extends true ? K : never]: OptionTypeMap[O[K]['type']]
 } & {
   [K in keyof O as IsRequired<O[K]> extends true ? never : K]?: OptionTypeMap[O[K]['type']]
-}
-
-/**
- * Returns a `CLIAdapter` as-is with full type inference. Pass a custom adapter to `createCLI` to swap the CLI engine.
- *
- * @example
- * ```ts
- * const adapter = defineCLIAdapter({ run: myRun, renderHelp: myHelp })
- * createCLI({ adapter }).run(commands, argv, opts)
- * ```
- */
-export function defineCLIAdapter(adapter: CLIAdapter): CLIAdapter {
-  return adapter
 }
 
 /**
