--- internals/utils/src/cli/schema.ts
@@ -1,45 +1,0 @@
-import type { CommandDefinition, CommandSchema, OptionDefinition, OptionSchema } from './types.ts'
-
-/**
- * Serializes `CommandDefinition[]` to a plain, JSON-serializable structure.
- * Use to expose CLI capabilities to AI agents or MCP tools.
- *
- * @example
- * ```ts
- * getCommandSchema([generateCommand, agentCommand])
- * // [{ name: 'generate', options: [...], subCommands: [] }, ...]
- * ```
- */
-export function getCommandSchema(defs: Array<CommandDefinition>): Array<CommandSchema> {
-  return defs.map(serializeCommand)
-}
-
-function serializeCommand(def: CommandDefinition): CommandSchema {
-  return {
-    name: def.name,
-    description: def.description,
-    arguments: def.arguments,
-    ...(def.examples?.length ? { examples: def.examples } : {}),
-    options: serializeOptions(def.options ?? {}),
-    subCommands: def.subCommands ? def.subCommands.map(serializeCommand) : [],
-  }
-}
-
-function serializeOptions(options: Record<string, OptionDefinition>): Array<OptionSchema> {
-  return Object.entries(options).map(([name, opt]) => {
-    const shortPart = opt.short ? `-${opt.short}, ` : ''
-    const valuePart = opt.type === 'string' ? ` <${opt.hint ?? name}>` : ''
-    const flags = `${shortPart}--${name}${valuePart}`
-
-    return {
-      name,
-      flags,
-      type: opt.type,
-      description: opt.description,
-      ...(opt.default !== undefined ? { default: opt.default } : {}),
-      ...(opt.hint ? { hint: opt.hint } : {}),
-      ...(opt.enum ? { enum: opt.enum } : {}),
-      ...(opt.required ? { required: opt.required } : {}),
-    }
-  })
-}
