+++ internals/utils/src/index.ts
@@ -1,15 +1,15 @@
 export { AsyncEventEmitter } from './asyncEventEmitter.ts'
 export { camelCase, pascalCase } from './casing.ts'
-export { defineCommand } from './cli/define.ts'
-export { createCLI } from './cli/parse.ts'
+export { createCLI } from './cli/createCLI.ts'
+export { defineCommand } from './cli/defineCommand.ts'
 export { formatMsWithColor, getIntro, randomCliColor } from './colors.ts'
 export { type BuildTree, buildTree } from './buildTree.ts'
 export { canUseTTY, isCIEnvironment } from './env.ts'
 export { BuildError, getErrorMessage, toCause, toError } from './errors.ts'
 export { clean, exists, findPackageJSON, getRelativePath, read, toFilePath, toPosixPath, trimExtName, write } from './fs.ts'
-export { isPlainObject, mergeDeep, serializePluginOptions } from './object.ts'
+export { isPlainObject, mergeDeep } from './object.ts'
 export type { ForBatchesOptions, PossiblePromise, Streamable } from './promise.ts'
-export { arrayToAsyncIterable, forBatches, isPromise, isPromiseRejectedResult, memoize } from './promise.ts'
+export { arrayToAsyncIterable, forBatches, isPromise, memoize } from './promise.ts'
 export { isIdentifier, isValidVarName, transformReservedWord } from './reserved.ts'
 export { runtime } from './runtime.ts'
 export { singleQuote } from './string.ts'
